Revert "Revert "refactor(suite-sync): rename reducer file to slice""
This reverts commit a6a454b6107992d292fe66915747f4878424071d.
