+++ suite-common/suite-sync/src/index.ts
@@ -14,8 +14,8 @@ export {
     updateSuiteSyncDebugEnabled,
     updateSuiteSyncEnabled,
     setSuiteSyncRelayUrl,
-} from './suiteSyncReducer';
-export type { SuiteSyncState, SuiteSyncSettings } from './suiteSyncReducer';
+} from './suiteSyncSlice';
+export type { SuiteSyncState, SuiteSyncSettings } from './suiteSyncSlice';
 export { DEFAULT_SUITE_SYNC_RELAY_URL } from './relay/relayUrl';
 export { prepareSuiteSyncMiddleware } from './suiteSyncMiddleware';
 export {
